Move to api/views/privaterelay.py
