+++ pyproject.toml
@@ -51,7 +51,7 @@ module = [
     "codetiming",
     "debug_toolbar",
     "dj_database_url",
-    "django_filters",
+    "django_filters.*",
     "django_ftl",
     "django_ftl.bundles",
     "google_measurement_protocol",
